feat: add ability to create annotation properties
Note, this is effectively syntactic sugar but it does make it a lot more consistent. Using a property definition simple prepend the new member name with '@' as per an annotation and the correct RDF types are generated for the property. WFR: may only use types that correspond to those legal for annotation property values.

```
module corpus_test_property is
  property @deprecatedSince -> {1} string is
    @skos:definition = "Used to denote when"@en
  end
end
```